(SIMP-1465) Safety save
========================================================================

       !!!   D O  N O T  M E R G E   T H I S   C O M M I T   !!!

========================================================================

This patch is the first stage of refactoring simp-rake-helpers into
testable classes and moving common code into other gems.

- A `Simp::Utils::Mock` class has been provided to run mock commands

SIMP-1465 #comment Added `Simp::Build::Mock` class + tests
SIMP-1465 #comment Updated `Simp::Rake::Pkg` to use `Simp::Build::Mock`
