+++ .gitignore
@@ -2,3 +2,7 @@
 Gemfile.lock
 dist/
 /log/
+*.log
+/doc/
+/.yardoc/
+/coverage/
