+++ .yardopts
@@ -1,0 +1,1 @@
+--markup markdown
