+++ bin/simp_rake_helpers
@@ -1,3 +1,7 @@
 #!/usr/bin/env ruby
+#!/usr/bin/env ruby
 
-abort "you need to write me"
+$LOAD_PATH << File.expand_path( '../lib', File.dirname(__FILE__) )
+require 'simp/build/cli'
+
+Simp::Build::Cli.start( ARGV )
