+++ simp-rake-helpers.gemspec
@@ -51,6 +51,8 @@ Gem::Specification.new do |s|
   s.add_development_dependency 'guard',       '~> 2.0'
   s.add_development_dependency 'guard-shell', '~> 0.0'
   s.add_development_dependency 'guard-rspec', '~> 4.0'
+  s.add_development_dependency 'yard',        '~> 0.9'
+  s.add_development_dependency 'simplecov'
 
 
   s.files = Dir[
